rename CapabilityRepository to CapabilityEntityFrameworkRepository & move CapabilityEntityFrameworkRepository to intrastructure
