+++ src/Harald.WebApi/Harald.WebApi.csproj
@@ -20,9 +20,4 @@
     <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Application" />
-    <Folder Include="Infrastructure\OutgoingAdapters" />
-  </ItemGroup>
-
 </Project>
